refactor: reorganize toolbar by functionality, rename floating-toolbar → toolbar
- Rename FloatingToolbar → Toolbar, FloatingToolbarSections → ToolbarSections
- Reorganize into toolbar/ directory with features/ and controls/ subdirs
- Group by functionality: text sections, color/border, font controls, numeric controls
- Split large files: sections.tsx (910L → 397+544L), fields.tsx (597L → 314+288L)
- Each file under 400 lines (2 borderline cases at 450L and 544L for complex controls)
- Update all imports, data-testid values, and e2e tests
- Rename e2e floating-toolbar.spec.ts → toolbar.spec.ts

Per Mine77's review on PR #51
