+++ src/editor/components/toolbar/controls/font-controls.tsx
@@ -1,21 +1,18 @@
 import { ChevronDown, Minus, Plus, Type } from "lucide-react";
 import { type KeyboardEvent as ReactKeyboardEvent, useEffect, useRef, useState } from "react";
-import type { ElementToolFeature } from "../lib/element-tool-model";
-import { FONT_FAMILY_OPTIONS, getFontFamilyLabel } from "../lib/style-controls";
-import { cn } from "../lib/utils";
-import { IconButton } from "./floating-toolbar-parts";
+import type { ElementToolFeature } from "../../../lib/element-tool-model";
+import { FONT_FAMILY_OPTIONS, getFontFamilyLabel } from "../../../lib/style-controls";
+import { cn } from "../../../lib/utils";
+import { IconButton } from "../primitives";
 import {
   ICON_STROKE_WIDTH,
   menuItemClassName,
   toolbarIconClassName,
   toolbarIconMutedClassName,
-} from "./floating-toolbar-types";
-import { Button } from "./ui/button";
-import { Input } from "./ui/input";
-import { Popover, PopoverAnchor, PopoverContent } from "./ui/popover";
-import { Textarea } from "./ui/textarea";
-
-const NUMBER_INPUT_DEBOUNCE_MS = 450;
+} from "../types";
+import { Button } from "../../ui/button";
+import { Input } from "../../ui/input";
+import { Popover, PopoverAnchor, PopoverContent } from "../../ui/popover";
 
 function FontFamilyCombobox({
   currentValue,
@@ -101,7 +98,7 @@ function FontFamilyCombobox({
             strokeWidth={ICON_STROKE_WIDTH}
           />
           <Input
-            aria-controls="floating-font-menu"
+            aria-controls="toolbar-font-menu"
             aria-expanded={open}
             aria-label="Font"
             autoComplete="off"
@@ -143,13 +140,13 @@ function FontFamilyCombobox({
         align="start"
         avoidCollisions
         className="w-[240px] p-1.5"
-        data-testid="floating-font-menu"
-        id="floating-font-menu"
+        data-testid="toolbar-font-menu"
+        id="toolbar-font-menu"
         onOpenAutoFocus={(event) => event.preventDefault()}
         side="bottom"
         sideOffset={18}
       >
-        <div className="h-56 overflow-y-auto pr-1" data-testid="floating-font-menu-scroll">
+        <div className="h-56 overflow-y-auto pr-1" data-testid="toolbar-font-menu-scroll">
           {filteredFonts.length ? (
             filteredFonts.map((font) => (
               <button
@@ -176,278 +173,6 @@ function FontFamilyCombobox({
         </div>
       </PopoverContent>
     </Popover>
-  );
-}
-
-function NumericCommitControl({
-  feature,
-  label,
-  onCommitFeature,
-  onPreview,
-  unit = "",
-}: {
-  feature: ElementToolFeature;
-  label: string;
-  onCommitFeature: (feature: ElementToolFeature, nextValue: string) => void;
-  onPreview?: (nextValue: string | null) => void;
-  unit?: string;
-}) {
-  const [draft, setDraft] = useState("");
-  const inputId = `floating-${feature.id}-custom`;
-
-  useEffect(() => {
-    const trimmedDraft = draft.trim();
-    if (!trimmedDraft || !onPreview) {
-      onPreview?.(null);
-      return;
-    }
-
-    const timeoutId = window.setTimeout(() => {
-      onPreview(`${trimmedDraft}${unit}`);
-    }, NUMERIC_INPUT_DEBOUNCE_MS);
-
-    return () => window.clearTimeout(timeoutId);
-  }, [draft, onPreview, unit]);
-
-  function commitDraft() {
-    const trimmedDraft = draft.trim();
-    if (!trimmedDraft) {
-      return;
-    }
-
-    onPreview?.(null);
-    onCommitFeature(feature, `${trimmedDraft}${unit}`);
-    setDraft("");
-  }
-
-  return (
-    <div className="grid gap-1.5">
-      <label
-        className="text-[10px] font-medium uppercase tracking-wider text-foreground/45"
-        htmlFor={inputId}
-      >
-        {label}
-      </label>
-      <div className="flex gap-1">
-        <Input
-          className="h-8 min-w-0"
-          id={inputId}
-          inputMode="decimal"
-          value={draft}
-          onChange={(event) => setDraft(event.target.value)}
-          onKeyDown={(event) => {
-            if (event.key === "Enter") {
-              event.preventDefault();
-              commitDraft();
-            }
-            if (event.key === "Escape") {
-              onPreview?.(null);
-              setDraft("");
-            }
-          }}
-        />
-        <div
-          className="grid h-8 min-w-11 place-items-center rounded-md border border-input bg-muted/30 px-2 text-[12px] font-medium text-foreground/45"
-          aria-label={unit ? `Unit ${unit}` : "Unitless value"}
-        >
-          {unit || "x"}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function DebouncedStyleNumberControl({
-  currentValue,
-  feature,
-  formatValue,
-  getDraftValue,
-  label,
-  max = 999,
-  min = 0,
-  onCommitFeature,
-  onPreviewStyle,
-  unit,
-}: {
-  currentValue: string;
-  feature: ElementToolFeature;
-  formatValue: (numericValue: number) => string;
-  getDraftValue: (currentValue: string) => string;
-  label: string;
-  max?: number;
-  min?: number;
-  onCommitFeature: (feature: ElementToolFeature, nextValue: string) => void;
-  onPreviewStyle: (propertyName: string, nextValue: string | null) => void;
-  unit: string;
-}) {
-  const [draft, setDraft] = useState(() => getDraftValue(currentValue));
-  const [isFocused, setIsFocused] = useState(false);
-  const hasPreviewedDraftRef = useRef(false);
-  const committedDraftRef = useRef(getDraftValue(currentValue));
-  const inputId = `floating-${feature.id}-custom`;
-
-  useEffect(() => {
-    const nextDraft = getDraftValue(currentValue);
-    committedDraftRef.current = nextDraft;
-    if (!isFocused) {
-      setDraft(nextDraft);
-    }
-  }, [currentValue, getDraftValue, isFocused]);
-
-  useEffect(() => {
-    return () => {
-      if (feature.propertyName && hasPreviewedDraftRef.current) {
-        onPreviewStyle(feature.propertyName, null);
-      }
-    };
-  }, [feature.propertyName, onPreviewStyle]);
-
-  useEffect(() => {
-    if (!isFocused || !draft.trim()) {
-      return;
-    }
-
-    const timeoutId = window.setTimeout(() => {
-      commitDraftValue(draft, false);
-    }, NUMBER_INPUT_DEBOUNCE_MS);
-
-    return () => window.clearTimeout(timeoutId);
-  }, [draft, isFocused]);
-
-  function getNextValue(nextDraft: string) {
-    const numericValue = Number.parseFloat(nextDraft);
-    if (!Number.isFinite(numericValue)) {
-      return null;
-    }
-
-    return formatValue(clamp(numericValue, min, max));
-  }
-
-  function previewDraftValue(nextDraft: string) {
-    if (!feature.propertyName) {
-      return;
-    }
-
-    const nextValue = getNextValue(nextDraft);
-    hasPreviewedDraftRef.current = nextValue !== null;
-    onPreviewStyle(feature.propertyName, nextValue);
-  }
-
-  function commitDraftValue(nextDraft: string, syncDraft: boolean) {
-    const trimmedDraft = nextDraft.trim();
-    if (!trimmedDraft) {
-      if (feature.propertyName) {
-        onPreviewStyle(feature.propertyName, null);
-      }
-      hasPreviewedDraftRef.current = false;
-      setDraft(getDraftValue(currentValue));
-      return;
-    }
-
-    const nextValue = getNextValue(trimmedDraft);
-    if (nextValue === null) {
-      setDraft(getDraftValue(currentValue));
-      return;
-    }
-
-    const normalizedDraft = String(clamp(Number.parseFloat(trimmedDraft), min, max));
-    if (feature.propertyName) {
-      onPreviewStyle(feature.propertyName, null);
-    }
-    hasPreviewedDraftRef.current = false;
-    committedDraftRef.current = normalizedDraft;
-    onCommitFeature(feature, nextValue);
-    if (syncDraft) {
-      setDraft(normalizedDraft);
-    }
-  }
-
-  return (
-    <div className="grid gap-1.5">
-      <label
-        className="text-[10px] font-medium uppercase tracking-wider text-foreground/45"
-        htmlFor={inputId}
-      >
-        {label}
-      </label>
-      <div className="flex h-8 items-center rounded-md border border-foreground/[0.08] bg-foreground/[0.03] px-2 transition-colors focus-within:border-foreground/20 focus-within:bg-white focus-within:ring-[2px] focus-within:ring-ring/35">
-        <Input
-          aria-label={label}
-          className="h-full min-w-0 border-0 bg-transparent px-0 text-[13px] shadow-none outline-none ring-0 [appearance:textfield] focus-visible:ring-0 [&::-webkit-inner-spin-button]:appearance-none [&::-webkit-outer-spin-button]:appearance-none"
-          id={inputId}
-          inputMode="decimal"
-          max={max}
-          min={min}
-          type="number"
-          value={draft}
-          onBlur={() => {
-            commitDraftValue(draft, true);
-            setIsFocused(false);
-          }}
-          onChange={(event) => {
-            const nextDraft = event.target.value;
-            setDraft(nextDraft);
-            previewDraftValue(nextDraft);
-          }}
-          onFocus={() => setIsFocused(true)}
-          onKeyDown={(event) => {
-            if (event.key === "Enter") {
-              event.currentTarget.blur();
-            }
-            if (event.key === "Escape") {
-              if (feature.propertyName) {
-                onPreviewStyle(feature.propertyName, null);
-              }
-              hasPreviewedDraftRef.current = false;
-              setDraft(getDraftValue(currentValue));
-              event.currentTarget.blur();
-            }
-          }}
-        />
-        <span
-          className="ml-2 shrink-0 text-[11px] font-medium text-foreground/45"
-          data-testid={`floating-${feature.id}-custom-unit`}
-        >
-          {unit}
-        </span>
-      </div>
-    </div>
-  );
-}
-
-function TextCommitControl({
-  feature,
-  label,
-  onCommitFeature,
-}: {
-  feature: ElementToolFeature;
-  label: string;
-  onCommitFeature: (feature: ElementToolFeature, nextValue: string) => void;
-}) {
-  const [draft, setDraft] = useState("");
-  const inputId = `floating-${feature.id}-custom`;
-  return (
-    <div className="grid gap-1.5">
-      <label
-        className="text-[10px] font-medium uppercase tracking-wider text-foreground/45"
-        htmlFor={inputId}
-      >
-        {label}
-      </label>
-      <Textarea id={inputId} value={draft} onChange={(event) => setDraft(event.target.value)} />
-      <Button
-        type="button"
-        variant="outline"
-        size="sm"
-        disabled={!draft.trim()}
-        onClick={() => {
-          onCommitFeature(feature, draft.trim());
-          setDraft("");
-        }}
-      >
-        Apply
-      </Button>
-    </div>
   );
 }
 
@@ -585,12 +310,5 @@ function parseFontSizeValue(value: string) {
 }
 
 const FONT_SIZE_INPUT_DEBOUNCE_MS = 500;
-const NUMERIC_INPUT_DEBOUNCE_MS = 500;
 
-export {
-  DebouncedStyleNumberControl,
-  FontFamilyCombobox,
-  FontSizeControl,
-  NumericCommitControl,
-  TextCommitControl,
-};
+export { FontFamilyCombobox, FontSizeControl };
