+++ src/editor/components/toolbar/index.ts
@@ -1,0 +1,2 @@
+export { Toolbar } from "./toolbar";
+export type { SelectionCommandAvailability } from "./types";
